test(python): share the echo server helper between the test modules
